refactor: reorganize eleventy configuration
