+++ artifacts/worklogs/20250910T210231Z.md
@@ -1,0 +1,4 @@
+# Worklog 20250910T210231Z
+- Attempted environment activation: scripts/llm-bootstrap.sh missing
+- Completed Eleventy build successfully
+- Attempted npm test: missing tools/runner.mjs
