+++ config/site.js
@@ -1,7 +1,5 @@
-/**
- * Global configuration shared across the build.
- * @module config
- */
+// config/site.js
+import path from 'node:path';
 
 /** Eleventy directory settings */
 export const dirs = {
@@ -11,4 +9,14 @@ export const dirs = {
   data: '_data',
 };
 
-export default { dirs };
+/** Root directory for markdown content */
+export const baseContentPath = 'src/content';
+
+/** Absolute path to concepts directory */
+export const CONCEPTS_DIR = path.join(baseContentPath, 'concepts');
+
+/** Primary content areas used for collections and navigation */
+export const CONTENT_AREAS = ['sparks', 'concepts', 'projects', 'archives', 'meta'];
+
+export default { dirs, baseContentPath, CONCEPTS_DIR, CONTENT_AREAS };
+
