+++ src/_data/archivesNav.js
@@ -1,3 +1,3 @@
 // Auto-synced: import navigation map from helpers
-import { buildArchiveNav } from '../../helpers/data/archive-nav.js';
+import { buildArchiveNav } from '../../utils/build/archive-nav.js';
 export default buildArchiveNav();
