+++ src/assets/css/mschf-overlay.css
@@ -1,0 +1,6 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+/* Minimal overlay style */
+.mschf-overlay-root { @apply relative block; }
